+++ ropa/spiders/benditopie.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
 import time
 from scrapy.http import Request
 from datetime import datetime
@@ -14,38 +16,16 @@ from pymongo import MongoClient
 
 from text_parser import price_normalize, html_text_normalize
 
-class BenditoPie(CrawlSpider):
+class BenditoPie(MioCrawler):
     name = 'benditopie'
     allowed_domains = ['benditopie.com']
 
     start_urls = ['https://benditopie.com/collections/zapatos-1?page=' + str(i) for i in range(1,7)]
                 
-
-
-    def __init__(self):
-        CrawlSpider.__init__(self)
-        self.verificationErrors = []
-        # self.browser = webdriver.PhantomJS()
-        self.browser = webdriver.Chrome()
-        self.browser.set_page_load_timeout(120)
-        self.connection = MongoClient("localhost", 27017)
-        self.comments = self.connection.ropa.items
-        self.links = self.connection.ropa.links
-
     rules = [
         # Rule(LinkExtractor(restrict_xpaths="//a[@class='f-linkNota']"), callback='parse_item', follow=True)
         # Rule(LinkExtractor(allow_domains=allowed_domains), callback='parse_item', follow=True)
     ]
-
-    def flaten_array_of_strings(self, array):
-        if len(array) > 0:
-            final_string = array[0]
-            for i in range(1, len(array)-1):
-                final_string += " " + array[i]
-            return(final_string)
-        else:
-            return("")
-
 
     def parse(self, response):
         print("------------- Crawling ----------------")
